Moved BaseSurface to base.py module
