+++ pvfactors/geometry/__init__.py
@@ -1,6 +1,6 @@
 """Geometry classes"""
 
-from pvfactors.geometry.base import BaseSide, ShadeCollection
+from pvfactors.geometry.base import \
+    BaseSide, ShadeCollection, BaseSurface
 from pvfactors.geometry.pvrow import PVRowSide, PVRow
-from pvfactors.geometry.pvsurface import \
-    BaseSurface, PVSurface, PVSegment
+from pvfactors.geometry.pvsurface import PVSurface, PVSegment
